+++ common/src/main/java/com/lne_archers/LNE_ArchersMod.java
@@ -9,7 +9,7 @@ import com.lne_archers.sounds.LneArchersSounds;
 import com.lne_archers.spells.CustomSpellImpacts;
 import com.lne_archers.config.TweaksConfig;
 import net.fabricmc.loader.api.FabricLoader;
-import net.spell_engine.api.config.ConfigFile;
+import net.spell_engine.rpg_series.config.ConfigFile;
 import net.tiny_config.ConfigManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;