+++ common/src/main/java/com/lne_archers/client/LNE_ArchersClient.java
@@ -7,7 +7,6 @@ import com.lne_archers.client.render.FrozenSlaveOverlayFeatureRenderer;
 import com.lne_archers.effects.Effects;
 import com.lne_archers.entity.InfiltratorsArrowProjectile;
 import com.lne_archers.entity.WintersGripEntity;
-import com.lne_archers.spells.ArchersSpells;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.LivingEntityFeatureRendererRegistrationCallback;
 import net.minecraft.client.render.entity.ArrowEntityRenderer;
@@ -15,15 +14,9 @@ import net.minecraft.client.render.entity.LivingEntityRenderer;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.entity.LivingEntity;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
-import net.spell_engine.client.gui.SpellTooltip;
 
 public class LNE_ArchersClient{
     public static void init() {
-        for (var entry : ArchersSpells.entries) {
-            if (entry.mutator() != null) {
-                SpellTooltip.addDescriptionMutator(entry.id(), entry.mutator());
-            }
-        }
         CustomParticleStatusEffect.register(Effects.RANGERS_FOCUS.effect, new RangersFocusParticles(15));
 
         EntityRendererRegistry.register(WintersGripEntity.ENTITY_TYPE, WintersGripRenderer::new);