+++ common/src/main/java/com/lne_archers/config/ItemConfig.java
@@ -1,7 +1,7 @@
 package com.lne_archers.config;
 
 import net.fabric_extras.ranged_weapon.api.RangedConfig;
-import net.spell_engine.api.config.WeaponConfig;
+import net.spell_engine.rpg_series.config.WeaponConfig;
 
 import java.util.LinkedHashMap;