+++ common/src/main/java/com/lne_archers/effects/WintersGraspInflicterAccess.java
@@ -1,0 +1,9 @@
+package com.lne_archers.effects;
+
+import java.util.UUID;
+
+public interface WintersGraspInflicterAccess {
+    UUID lneArchers$getWintersGraspInflicter();
+
+    void lneArchers$setWintersGraspInflicter(UUID inflicter);
+}